+++ src/main/java/lotterymachine/LotteryMachineApplication.java
@@ -6,7 +6,7 @@ import lotterymachine.domain.WinningLottery;
 import lotterymachine.domain.WinningResult;
 import lotterymachine.domain.vo.Count;
 import lotterymachine.domain.vo.Money;
-import lotterymachine.utils.RandomLotteryNumbersGenerator;
+import lotterymachine.domain.RandomLotteryNumbersGenerator;
 import lotterymachine.view.InputView;
 import lotterymachine.view.OutputView;
