+++ src/main/java/lotterymachine/domain/LotteryNumbersGenerator.java
@@ -1,6 +1,4 @@
-package lotterymachine.utils;
-
-import lotterymachine.domain.LotteryNumber;
+package lotterymachine.domain;
 
 import java.util.List;
